feat: add inbound funding-token transfer helper with balance-delta checks and tests
