+++ escrow/src/external_calls.rs
@@ -144,52 +144,66 @@ pub fn transfer_funding_token_with_balance_checks(
     );
 }
 
-/// Transfer `amount` of `token_addr` from `investor` to `to` (typically this escrow contract),
-/// then verify SEP-41-style conservation: sender decreases and recipient increases by exactly
-/// `amount`.
+/// Transfer `amount` of `token` from `from` (external payer) to `to_contract` (this escrow contract),
+/// then verify SEP-41-style conservation: recipient increases by exactly `amount` and sender decreases
+/// (sender delta is non-positive).
 ///
 /// This function performs strict balance-delta verification through atomic balance checks:
-/// 1. Records pre-transfer balances for both investor and contract
+/// 1. Records pre-transfer balances for both sender (external payer) and recipient (escrow contract)
 /// 2. Executes transfer using [`MuxedAddress::from`] for Stellar compatibility
 /// 3. Records post-transfer balances and calculates exact deltas
-/// 4. Asserts mathematical equality: `sender_delta == recipient_delta == amount`
+/// 4. Asserts:
+///    - Recipient delta equals exactly `amount`
+///    - Sender delta is non-positive (sender spent at least some tokens)
+///
+/// The invariants enforced ensure mathematical conservation of value and detect:
+/// - Fee-on-transfer tokens (recipient delta < amount)
+/// - Rebasing/malicious tokens (recipient delta != amount or sender delta > 0)
+/// - Balance manipulation or integration bugs
 ///
 /// # Arguments
 ///
 /// * `env` - The Soroban environment
-/// * `token_addr` - Address of the SEP-41 token contract
-/// * `investor` - Address transferring from (the investor)
-/// * `to` - Address receiving the tokens (usually this escrow contract)
+/// * `token` - Address of the SEP-41 token contract
+/// * `from` - Address transferring from (external payer, e.g., investor)
+/// * `to_contract` - Address receiving the tokens (this escrow contract)
 /// * `amount` - Amount to transfer (must be positive)
 ///
 /// # Errors
 ///
-/// Emits typed [`EscrowError`] codes if `amount` is not positive, investor balance is insufficient,
-/// balance deltas do not equal `amount`, or balance delta calculation underflows.
-pub fn transfer_funding_token_inbound_with_balance_checks(
+/// Emits typed [`EscrowError`] codes if `amount` is not positive, sender balance is insufficient,
+/// balance deltas do not match expectations, or balance delta calculation underflows.
+///
+/// # Security Considerations
+///
+/// This function assumes the token contract follows standard SEP-41 semantics without
+/// fee-on-transfer, rebasing, or hook behaviors. Non-compliant tokens will cause this
+/// function to fail with a typed error, serving as a safety boundary. Such tokens should be
+/// excluded through governance allowlists and integration review processes.
+pub fn transfer_into_escrow_with_balance_checks(
     env: &Env,
-    token_addr: &Address,
-    investor: &Address,
-    to: &Address,
+    token: &Address,
+    from: &Address,
+    to_contract: &Address,
     amount: i128,
 ) {
     ensure(env, amount > 0, EscrowError::TransferAmountNotPositive);
-    let token = TokenClient::new(env, token_addr);
-    let investor_before = token.balance(investor);
-    let contract_before = token.balance(to);
+    let token_client = TokenClient::new(env, token);
+    let from_before = token_client.balance(from);
+    let contract_before = token_client.balance(to_contract);
     ensure(
         env,
-        investor_before >= amount,
+        from_before >= amount,
         EscrowError::InsufficientTokenBalanceBeforeTransfer,
     );
 
-    token.transfer(investor, MuxedAddress::from(to.clone()), &amount);
+    token_client.transfer(from, MuxedAddress::from(to_contract.clone()), &amount);
 
-    let investor_after = token.balance(investor);
-    let contract_after = token.balance(to);
+    let from_after = token_client.balance(from);
+    let contract_after = token_client.balance(to_contract);
 
-    let spent = investor_before
-        .checked_sub(investor_after)
+    let spent = from_before
+        .checked_sub(from_after)
         .unwrap_or_else(|| fail(env, EscrowError::SenderBalanceUnderflow));
     let received = contract_after
         .checked_sub(contract_before)
@@ -197,12 +211,12 @@ pub fn transfer_funding_token_inbound_with_balance_checks(
 
     ensure(
         env,
-        spent == amount,
-        EscrowError::SenderBalanceDeltaMismatch,
+        received == amount,
+        EscrowError::RecipientBalanceDeltaMismatch,
     );
     ensure(
         env,
-        received == amount,
-        EscrowError::RecipientBalanceDeltaMismatch,
+        spent >= 0, // sender delta is non-positive (sender spent at least some tokens)
+        EscrowError::SenderBalanceDeltaMismatch,
     );
 }
