+++ escrow/src/tests/external_calls_mocked.rs
@@ -3,7 +3,7 @@
 //! This module contains tests that would fail if balance deltas diverge from expected behavior.
 //! Uses mocked token implementations where feasible in the Soroban test harness.
 
-use super::super::external_calls::transfer_funding_token_with_balance_checks;
+use super::super::external_calls::{transfer_funding_token_with_balance_checks, transfer_into_escrow_with_balance_checks};
 use super::*;
 use soroban_sdk::{contract, contractimpl, token::TokenInterface, Address, Env, MuxedAddress};
 // ---------------------------------------------------------------------------
@@ -575,4 +575,112 @@ fn test_insufficient_balance_panics_with_insufficient_token_balance() {
     // Mint nothing - balance is 0
     // Error code 37: InsufficientTokenBalanceBeforeTransfer
     transfer_funding_token_with_balance_checks(&env, &token.id, &holder, &treasury, 1);
+}
+
+// ---------------------------------------------------------------------------
+// Tests: inbound transfer_into_escrow_with_balance_checks
+// ---------------------------------------------------------------------------
+
+#[test]
+#[should_panic]
+fn test_inbound_fee_on_transfer_token_rejected() {
+    let env = Env::default();
+    env.mock_all_auths();
+    let fee_token_id = env.register(FeeOnTransferToken, ());
+    let investor = Address::generate(&env);
+    let escrow = deploy_id(&env);
+    mint_fee_token(&env, &fee_token_id, &investor, 1000i128);
+    // Recipient (escrow) receives less than amount -> panic
+    transfer_into_escrow_with_balance_checks(&env, &fee_token_id, &investor, &escrow, 1000i128);
+}
+
+#[test]
+#[should_panic]
+fn test_inbound_zero_amount_rejected() {
+    let env = Env::default();
+    env.mock_all_auths();
+    let token = install_stellar_asset_token(&env);
+    let investor = deploy_id(&env);
+    let escrow = Address::generate(&env);
+    transfer_into_escrow_with_balance_checks(&env, &token.id, &investor, &escrow, 0);
+}
+
+#[test]
+#[should_panic]
+fn test_inbound_negative_amount_rejected() {
+    let env = Env::default();
+    env.mock_all_auths();
+    let token = install_stellar_asset_token(&env);
+    let investor = deploy_id(&env);
+    let escrow = Address::generate(&env);
+    transfer_into_escrow_with_balance_checks(&env, &token.id, &investor, &escrow, -1i128);
+}
+
+#[test]
+#[should_panic]
+fn test_inbound_insufficient_balance_rejected() {
+    let env = Env::default();
+    env.mock_all_auths();
+    let token = install_stellar_asset_token(&env);
+    let investor = deploy_id(&env);
+    let escrow = Address::generate(&env);
+    // Investor has no tokens
+    transfer_into_escrow_with_balance_checks(&env, &token.id, &investor, &escrow, 1i128);
+}
+
+#[test]
+#[should_panic]
+fn test_inbound_lying_token_no_change_rejected() {
+    let env = Env::default();
+    env.mock_all_auths();
+    let lying_token_id = env.register(LyingToken, ());
+    let investor = Address::generate(&env);
+    let escrow = deploy_id(&env);
+    mint_lying_token(&env, &lying_token_id, &investor, 1000i128);
+    // No balance change -> RecipientBalanceDeltaMismatch
+    transfer_into_escrow_with_balance_checks(&env, &lying_token_id, &investor, &escrow, 1000i128);
+}
+
+#[test]
+#[should_panic]
+fn test_inbound_hook_token_recipient_decreases_rejected() {
+    let env = Env::default();
+    env.mock_all_auths();
+    let hook_token_id = env.register(HookStealingToken, ());
+    let investor = Address::generate(&env);
+    let escrow = deploy_id(&env);
+    mint_hook_token(&env, &hook_token_id, &investor, 1000i128);
+    // Hook reduces escrow balance after transfer
+    transfer_into_escrow_with_balance_checks(&env, &hook_token_id, &investor, &escrow, 1000i128);
+}
+
+#[test]
+#[should_panic]
+fn test_inbound_rebasing_token_sender_increases_rejected() {
+    let env = Env::default();
+    env.mock_all_auths();
+    let rebase_token_id = env.register(RebasingToken, ());
+    let investor = Address::generate(&env);
+    let escrow = deploy_id(&env);
+    mint_rebasing_token(&env, &rebase_token_id, &investor, 1000i128);
+    // Sender ends with extra tokens -> SenderBalanceDeltaMismatch
+    transfer_into_escrow_with_balance_checks(&env, &rebase_token_id, &investor, &escrow, 1000i128);
+}
+
+#[test]
+fn test_inbound_compliant_token_passes() {
+    let env = Env::default();
+    env.mock_all_auths();
+    let token = install_stellar_asset_token(&env);
+    let investor = deploy_id(&env);
+    let escrow = Address::generate(&env);
+    let amount = 1000i128;
+    token.stellar.mint(&investor, &amount);
+    let investor_before = token.token.balance(&investor);
+    let escrow_before = token.token.balance(&escrow);
+    transfer_into_escrow_with_balance_checks(&env, &token.id, &investor, &escrow, amount);
+    let investor_after = token.token.balance(&investor);
+    let escrow_after = token.token.balance(&escrow);
+    assert_eq!(investor_before - investor_after, amount);
+    assert_eq!(escrow_after - escrow_before, amount);
 }
